Linear Regression
In the activity
I tried changing the test value for page speed and noted the variations in r value. As the value, I increased of standard deviation by 0.5 value the number started coming closer to the 1. When mu=3.0, sigma=6.0 and x=1000 the r value we received is 0.99 which is nearly 1.0 it becomes a good fit